+++ R/spread.R
@@ -18,11 +18,11 @@
 #' # Spread and gather are complements
 #' df <- data.frame(x = c("a", "b"), y = c(3, 4), z = c(5, 6))
 #' df %>% spread(x, y) %>% gather(x, y, a:b, na.rm = TRUE)
-spread <- function(data, key, value, fill = NA, convert = FALSE) {
+spread <- function(data, key, value, fill = NA, convert = FALSE, drop = TRUE) {
   key_col <- col_name(substitute(key))
   value_col <- col_name(substitute(value))
 
-  spread_(data, key_col, value_col, fill = fill, convert = convert)
+  spread_(data, key_col, value_col, fill = fill, convert = convert, drop = drop)
 }
 
 #' Standard-evaluation version of \code{spread}.
@@ -38,21 +38,22 @@ spread <- function(data, key, value, fill = NA, convert = FALSE) {
 #'   useful if the value column was a mix of variables that was coerced to
 #'   a string.
 #' @export
-spread_ <- function(data, key_col, value_col, fill = NA, convert = FALSE) {
+spread_ <- function(data, key_col, value_col, fill = NA, convert = FALSE,
+                    drop = TRUE) {
   UseMethod("spread_")
 }
 
 #' @export
 spread_.data.frame <- function(data, key_col, value_col, fill = NA,
-                               convert = FALSE) {
+                               convert = FALSE, drop = TRUE) {
 
   col <- data[key_col]
-  col_id <- dplyr::id(col, drop = TRUE)
-  col_labels <- col[match(sort(unique(col_id)), col_id), , drop = FALSE]
+  col_id <- dplyr::id(col, drop = drop)
+  col_labels <- split_labels(col, col_id, drop = drop)
 
   rows <- data[setdiff(names(data), c(key_col, value_col))]
-  row_id <- dplyr::id(rows, drop = TRUE)
-  row_labels <- rows[match(sort(unique(row_id)), row_id), , drop = FALSE]
+  row_id <- dplyr::id(rows, drop = drop)
+  row_labels <- split_labels(rows, row_id, drop = drop)
   rownames(row_labels) <- NULL
 
   overall <- dplyr::id(list(col_id, row_id), drop = FALSE)
@@ -89,7 +90,6 @@ spread_.data.frame <- function(data, key_col, value_col, fill = NA,
   if (convert) {
     ordered[] <- lapply(ordered, type.convert, as.is = TRUE)
   }
-
   append_df(row_labels, ordered)
 }
 
@@ -97,4 +97,28 @@ spread_.data.frame <- function(data, key_col, value_col, fill = NA,
 spread_.tbl_df <- function(data, key_col, value_col, fill = NA,
                            convert = FALSE) {
   dplyr::tbl_df(NextMethod())
+}
+
+split_labels <- function(df, id, drop = TRUE) {
+  if (length(df) == 0) {
+    return(df)
+  }
+
+  if (drop) {
+    representative <- match(sort(unique(id)), id)
+    df[representative, , drop = FALSE]
+  } else {
+    unique_values <- lapply(df, ulevels)
+    rev(expand.grid(rev(unique_values), stringsAsFactors = FALSE))
+  }
+}
+
+ulevels <- function(x) {
+  if (is.factor(x)) {
+    x <- addNA(x, ifany = TRUE)
+    levs <- levels(x)
+    factor(levs, levels = levs)
+  } else {
+    sort(unique(x))
+  }
 }